requirements.txt
dependency-injector~=4.37.0
fastapi~=0.70.0
httpx~=0.20.0
mypy~=0.910
orjson~=3.6.4
pydantic~=1.8.2
pylint~=2.11.1
tortoise-orm[asyncpg]~=0.17.8
uvicorn[standard]~=0.15.0
